+++ CHANGELOG.md
@@ -13,6 +13,8 @@ All notable changes to this library are documented in this file.
 
 ### Bug Fixes
 
+- Account for a corrupted HRRR bufkit file found in MTArchive.
+
 ## **1.25.0** (30 Sep 2025)
 
 ### API Changes